pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "poiidx"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Janna Hopp" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "psycopg2-binary",
  "peewee",
  "shapely",
  "requests",
  "pyyaml",
  "platformdirs",
  "osmium",
  "pyproj",
]

[project.urls]
Documentation = "https://github.com/bytehexe/poiidx#readme"
Issues = "https://github.com/bytehexe/poiidx/issues"
Source = "https://github.com/bytehexe/poiidx"

[tool.hatch.version]
path = "src/poiidx/__about__.py"

[tool.hatch.envs.default]
extra-dependencies = [
  "click"
]

[tool.hatch.envs.default.scripts]
example = "python -m poiidx.example {args}"

[tool.hatch.envs.hatch-test]
extra-dependencies = [
  "pytest>=7.0.0",
  "testcontainers",
]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/poiidx tests}"

[tool.coverage.run]
source_pkgs = ["poiidx", "tests"]
branch = true
parallel = true
omit = [
  "src/poiidx/__about__.py",
]

[tool.coverage.paths]
poiidx = ["src/poiidx", "*/poiidx/src/poiidx"]
tests = ["tests", "*/poiidx/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
[tool.pytest.ini_options]
filterwarnings = [
  "ignore::DeprecationWarning:testcontainers.*",
]